# [stdlib] Speed up Character.init significantly for small characters.