@slavapestov 2017-01-17
It would be great to add a new benchmark for this!