@xwu 2017-01-17
One more Q: Now that you bypass the preconditions in `init(_: String)`, how is it ensured that you're storing only the first character in `.small`? If there is more than one character stored, I think `String.init(_: Character)` will spit the subsequent ones back out...