@allevato 2017-01-17
@slavapestov I'll look into adding one!

@xwu AFAICT, the compiler won't generate a call to this initializer for a literal that contains more than one grapheme cluster because it would be caught as an error before then. If you try to write `"ab" as Character` today, you get the error `cannot convert value of type 'String' to type 'Character' in coercion`, which I believe is enforced based on the fact that `Character` conforms to `ExpressibleByExtendedGraphemeClusterLiteral` rather than `ExpressibleByStringLiteral`. `init(_: String)` needs the check because the string received at runtime could contain any number of characters.