@airspeedswift 2017-01-17
@swift-ci Please benchmark