@allevato 2017-01-17
@airspeedswift Sounds good—I'll add some new benchmarks as soon as I have time to work on it again in the next couple of days and send a separate PR for those.