@allevato 2017-01-18
Now that #6879 is merged, do I need to do anything extra to re-run the benchmarks? (When I tried running it on that PR, nothing happened—are the swift-ci commands only available to project collaborators?)