@gottesmm 2017-01-18
@swift-ci Please benchmark