@airspeedswift 2017-01-19
@swift-ci Please smoke test linux platform