@allevato 2017-01-19
Wonderful! Those improvements jibe with what I observed locally.

Are the regressions shown there something I should look into, or is it statistical noise/within tolerance?