@slavapestov 2017-01-24
@allevato We can always just run the benchmarks again :-)