@slavapestov 2017-01-24
@swift-ci Please benchmark