@airspeedswift 2017-01-24
Yeah the second bench run suggests the regressions we are seeing are just jitter, not caused by the change.

I think we can merge this. @dabrahams do you agree?